名片信息绑定，添加选择相机相册功能

+++ lib/models/cardmodel.dart
@@ -27,7 +27,7 @@ class CardDataModel {
   String photo;
   String weChat;
   int customerCount;
-  int manageCapital;
+  double manageCapital;
   String company;
   String companyAddress;
   String position;
